move QueryResult to shardingsphere executor module
