+++ shardingsphere-underlying/shardingsphere-merge/src/test/java/org/apache/shardingsphere/underlying/merge/result/impl/memory/MemoryQueryResultRowTest.java
@@ -17,7 +17,7 @@
 
 package org.apache.shardingsphere.underlying.merge.result.impl.memory;
 
-import org.apache.shardingsphere.underlying.executor.QueryResult;
+import org.apache.shardingsphere.underlying.executor.sql.queryresult.QueryResult;
 import org.junit.Before;
 import org.junit.Test;
