+++ shardingsphere-underlying/shardingsphere-merge/src/test/java/org/apache/shardingsphere/underlying/merge/result/impl/memory/fixture/TestMemoryMergedResult.java
@@ -20,7 +20,7 @@ package org.apache.shardingsphere.underlying.merge.result.impl.memory.fixture;
 import lombok.Getter;
 import org.apache.shardingsphere.sql.parser.binder.metadata.schema.SchemaMetaData;
 import org.apache.shardingsphere.sql.parser.binder.statement.SQLStatementContext;
-import org.apache.shardingsphere.underlying.executor.QueryResult;
+import org.apache.shardingsphere.underlying.executor.sql.queryresult.QueryResult;
 import org.apache.shardingsphere.underlying.merge.result.impl.fixture.TestRule;
 import org.apache.shardingsphere.underlying.merge.result.impl.memory.MemoryMergedResult;
 import org.apache.shardingsphere.underlying.merge.result.impl.memory.MemoryQueryResultRow;
